backend/app/services/auth_service.py
from datetime import datetime, timedelta
from typing import Optional
from jose import JWTError, jwt
from passlib.context import CryptContext
from app.services.excel_service import read_sheet, write_sheet
from app.models.schemas import TokenData

# SECRET_KEY should be in env but hardcoding for now as per instructions/simplicity
SECRET_KEY = "09d25e094faa6ca2556c818166b7a9563b93f7099f6f0f4caa6cf63b88e8d3e7"
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 30

pwd_context = CryptContext(schemes=["pbkdf2_sha256"], deprecated="auto")

def verify_password(plain_password, hashed_password):
    return pwd_context.verify(plain_password, hashed_password)

def get_password_hash(password):
    return pwd_context.hash(password)

def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=15)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt

def authenticate_master(email: str, password: str):
    users = read_sheet("master", "users_master")
    user = next((u for u in users if u["email"] == email), None)
    
    if not user:
        # Fallback for hardcoded check if excel fail, or just rely on excel
        # Init script added the user, so it should be there.
        return False
        
    # In master excel we don't store password hash?
    # Request said: "Pode ter um usuário master definido em código".
    # But later "password_hash (hash bcrypt da senha do usuário)" listed for tenant users.
    # For master users in excel, we assume they might not have password there or simple.
    # Let's check if the excel row has password_hash, otherwise use hardcoded for the specific email.
    
    if email == "master@seucrm.com":
         # Hardcoded password check as per prompt suggestion for simplicity
         if verify_password(password, pwd_context.hash("Teste@123")):
             return user
    
    # If we added password_hash to master excel, check it here.
    # The init script did not add password col to master excel definition.
    return False

def get_tenant_file(tenant_slug: str):
    ambientes = read_sheet("ambientes", "ambientes")
    ambiente = next((a for a in ambientes if a["slug"] == tenant_slug and a["ativo"]), None)
    if ambiente:
        return ambiente["excel_file"]
    return None

def authenticate_tenant_user(tenant_slug: str, email: str, password: str):
    excel_file = get_tenant_file(tenant_slug)
    if not excel_file:
        return None
        
    users = read_sheet(excel_file.replace(".xlsx", ""), "users")
    user = next((u for u in users if u["email"] == email), None)
    
    if not user:
        return None
        
    if not verify_password(password, user["password_hash"]):
        return None
        
    return user

def change_user_password(email: str, old_password: str, new_password: str, role_global: str = None, tenant_slug: str = None):
    # Determine which file to look in
    if role_global == "master" and not tenant_slug:
        file_name = "master"
        sheet_name = "users_master"
    elif tenant_slug:
        # Tenant user or Master acting as Tenant Admin (though Master usually changes own pass in Master context)
        # If Master is logged in as Master but in Tenant Context, we probably shouldn't let them change the *Tenant User* password via this endpoint 
        # unless we explicitly allow Admin to reset user passwords (different feature).
        # This function is for "Change My Password".
        # So we look at who the user IS.
        
        # If role_global is master, their record is in master.xlsx, even if they are viewing a tenant.
        if role_global == "master":
             file_name = "master"
             sheet_name = "users_master"
        else:
             from app.services.auth_service import get_tenant_file
             excel_file = get_tenant_file(tenant_slug)
             if not excel_file: return False
             file_name = excel_file.replace(".xlsx", "")
             sheet_name = "users"
    else:
        return False

    users = read_sheet(file_name, sheet_name)
    user_idx = next((i for i, u in enumerate(users) if u["email"] == email), -1)
    
    if user_idx == -1:
        return False
    
    user = users[user_idx]
    
    # Verify old password
    # Master might not have password_hash field in schema if simplified?
    # Logic in authenticate_master checked hardcoded.
    # To support change password, we MUST write the hash back.
    # If legacy master in code, we can't change it here easily unless we migrate to excel flow completely.
    # Let's assume we migrated or will overwrite.
    
    current_hash = user.get("password_hash")
    # If no hash (legacy hardcoded), checks might fail if we don't handle it.
    # For now, let's assume standard behavior.
    
    if not current_hash:
        # Fallback for the hardcoded "Teste@123" if user matches
        if email == "master@seucrm.com" and verify_password(old_password, pwd_context.hash("Teste@123")):
             # Allow change
             pass
        else:
             return False
    elif not verify_password(old_password, current_hash):
        return False
        
    # Update
    user["password_hash"] = get_password_hash(new_password)
    users[user_idx] = user
    
    write_sheet(file_name, sheet_name, users)
    return True
