backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
openpyxl
pandas
python-multipart
fpdf2
